Refactor: Moved menu creation to createMainWindow. Not needed in main
