+++ src/index.ts
@@ -1,20 +1,14 @@
 import {
   app,
   BrowserWindow,
-  dialog,
-  globalShortcut,
-  Menu,
-  type MenuItem
+  dialog
 } from 'electron'
 import * as log from 'electron-log'
-import { appMenuTemplate, setMenuBarVisibility } from './menu'
 import { conf } from './appConfig'
 import { createMainWindow } from './mainWindow'
 import { createAboutWindow } from './aboutWindow'
 import { createTrayItem } from './trayItem'
 // TODO: Implement tray item, tray context menu, open from tray, etc
-
-// global reference for main window object
 
 const currentVersion = app.getVersion()
 conf.set('currentVersion', currentVersion)
@@ -46,24 +40,6 @@ app.whenReady()
         mainWindow.show()
       }
     })
-
-    /* ******************************************************************** */
-    // create menu
-    const menuTemplate = appMenuTemplate(conf)
-    const menu = Menu.buildFromTemplate(menuTemplate)
-    Menu.setApplicationMenu(menu)
-
-    // modify menu bar
-    setMenuBarVisibility(mainWindow, !conf.get('autoHideMenuBar'))
-    // register global shortcuts
-    globalShortcut.register('F1', () => {
-      createAboutWindow(mainWindow) // show about window on F1 keypress
-    })
-    // set menu checkbox values
-    setMenuCheckbox(menu, 'autoHideMenuBar')
-    setMenuCheckbox(menu, 'minimizeToTray')
-    setMenuCheckbox(menu, 'closeToTray')
-    setMenuCheckbox(menu, 'enableGPUAcceleration')
   })
   .catch((err) => {
     log.error(err)
@@ -77,10 +53,3 @@ app.on('window-all-closed', () => {
     app.quit()
   }
 })
-/* ******************************************************************** */
-// menu checkbox functions
-// app menu generic helper functions
-const setMenuCheckbox = (menu: Menu, menuItemId: string): void => {
-  const menuItem = menu.getMenuItemById(menuItemId) as MenuItem
-  menuItem.checked = conf.get(menuItemId)
-}
