+++ src/mainWindow.ts
@@ -1,7 +1,9 @@
 import path from 'path'
-import { app, BrowserWindow, Menu, type MenuItem, shell } from 'electron'
+import { app, BrowserWindow, globalShortcut, Menu, type MenuItem, shell } from 'electron'
 import * as log from 'electron-log'
 import { conf } from './appConfig'
+import { appMenuTemplate, setMenuBarVisibility } from './menu'
+import { createAboutWindow } from './aboutWindow'
 
 const currentVersion = app.getVersion()
 export const createMainWindow = (): BrowserWindow => {
@@ -87,5 +89,29 @@ export const createMainWindow = (): BrowserWindow => {
     return { action: 'deny' }
   })
 
+  /* ******************************************************************** */
+  // create app menu
+  const menuTemplate = appMenuTemplate(conf)
+  const menu = Menu.buildFromTemplate(menuTemplate)
+  Menu.setApplicationMenu(menu)
+
+  // modify menu bar
+  setMenuBarVisibility(win, !conf.get('autoHideMenuBar'))
+  // register global shortcuts
+  globalShortcut.register('F1', () => {
+    createAboutWindow(win) // show about window on F1 keypress
+  })
+  // set menu checkbox values
+  setMenuCheckbox(menu, 'autoHideMenuBar')
+  setMenuCheckbox(menu, 'minimizeToTray')
+  setMenuCheckbox(menu, 'closeToTray')
+  setMenuCheckbox(menu, 'enableGPUAcceleration')
+
   return win
+}
+
+// menu checkbox helper function
+const setMenuCheckbox = (menu: Menu, menuItemId: string): void => {
+  const menuItem = menu.getMenuItemById(menuItemId) as MenuItem
+  menuItem.checked = conf.get(menuItemId)
 }
